[rhio config] minor fixes
