+++ rhio-http-api/src/blocking.rs
@@ -4,6 +4,22 @@ use crate::{api::RhioApi, status::HealthStatus};
 use anyhow::Result;
 use tokio::runtime::Runtime;
 
+/// A blocking client for interacting with the Rhio HTTP API.
+///
+/// The `BlockingClient` struct provides a synchronous interface to the asynchronous
+/// `RhioApi` by using a Tokio runtime to block on asynchronous calls. This allows
+/// users to interact with the API in a blocking manner, which can be useful in
+/// contexts where asynchronous code is not suitable.
+///
+/// # Type Parameters
+///
+/// * `A` - A type that implements the `RhioApi` trait.
+///
+/// # Fields
+///
+/// * `inner` - The inner API client that implements the `RhioApi` trait.
+/// * `runtime` - An `Arc` wrapped Tokio runtime used to block on asynchronous calls.
+///
 pub struct BlockingClient<A>
 where
     A: RhioApi,
